+++ typescript/src/index.ts
@@ -1,7 +1,8 @@
 //Services
-export { StaticFileID, FilesArtifactID, ContainerConfig, ContainerConfigBuilder } from "./lib/services/container_config"; 
+export { FilesArtifactID, ContainerConfig, ContainerConfigBuilder } from "./lib/services/container_config";
 export { ServiceID } from "./lib/services/service";
 export { ServiceContext } from "./lib/services/service_context";
+export { SharedPath } from "./lib/services/shared_path"
 
 //Networks
 export { Network } from "./lib/networks/network";