+++ typescript/src/lib/services/container_config.ts
@@ -1,5 +1,3 @@
-export type StaticFileID = string;
-
 // The ID of an artifact containing files that should be mounted into a service container
 export type FilesArtifactID = string;