+++ docs/changelog.md
@@ -1,6 +1,7 @@
 # TBD
 ### Features
-* Add `SharedDirectory` object which contains two fields `absFilepathOnThisContainer` and `absFilepathOnServiceContainer` that store the values of the service directory absolute path in the testsuite container and the service container
+* Add `SharedPath` object which contains two fields `absPathOnThisContainer` and `absPathOnServiceContainer` that store the values of an absolute path in the container where the code is running and in the service container
+* Add `SharedPath.GetChildPath()` which returns a new `SharedPath` composed by the actual value and adding it a new path element at the end of it
 * Update API container service proto definition adding the `relative_service_dirpath` field in `RegisterServiceResponse` and `GetServiceInfoResponse`
 
 ### Changes
@@ -11,8 +12,8 @@
   * Users should replace the creation of `ContainerCreationConfig` and `ContainerRunConfig` objects with an implementation of an anonymous function that dynamically generates the new `ContainerConfig` object
 * Update `NetworkCtx.AddService()` method replacing the `containerCreationConfig` param and the `generateRunConfigFunc` with the `containerConfigSupplier` param
   * Users should update `NetworkCtx.AddService()` calls passing it now the `containerConfigSupplier` which is an anonymous function that should be created to dynamically generate the `containerConfig` object 
-* Replace `enclaveDataVolMountpointHere` and `enclaveDataVolMountpointOnServiceContainer` in `ServiceContext` fields with a `SharedDirectory` object
-  * Users should use the fields of `SharedDirectory` object to get the service directory absolute paths
+* Replace `enclaveDataVolMountpointHere` and `enclaveDataVolMountpointOnServiceContainer` in `ServiceContext` fields with a new field called `sharedDirectory` with `SharedPath` type
+  * Users should use the `SharedPath` object store id the `sharedDirectory` field to get the service directory absolute paths
 * Removed the `ServiceContext.GenerateFiles()` and the `ServiceContext.LoadStaticFiles()` methods
   * Users should manually create, generate and copy static and dynamic files into the service container with the help of the new `SharedDirectory` object